+++ src/Gitle.Model/InvoiceLine.cs
@@ -16,7 +16,7 @@ namespace Gitle.Model
         public virtual Issue Issue { get; set; }
         public virtual string Description { get; set; }
         public virtual double Hours { get; set; }
-        public virtual double Price => Null ? 0 : Hours * Invoice.HourPrice;
+        public virtual decimal Price => Null ? 0 : (decimal)Hours * Invoice.HourPrice;
 
         public virtual bool Null
         {
@@ -44,6 +44,6 @@ namespace Gitle.Model
     {
         public virtual Invoice Invoice { get; set; }
         public virtual string Description { get; set; }
-        public virtual double Price { get; set; }
+        public virtual decimal Price { get; set; }
     }
 }
